requirements.txt
apache-airflow==3.1.3
pandas==2.1.3
pydantic==2.5.0
requests==2.31.0
typing-extensions==4.9.0
cryptography==41.0.7